Adjust the YSP scenario to the semantics
According to the semantics,
if the domain description contains an "f triggers Action" statement,
then, at every timepoint f is true, Action is executed by some agent.

Make the triggered action ("escape") make the trigger condition false,
so that it's not triggered forever
and a "shoot" action can be executed later.